requirements.txt
langchain==0.2.14
langchain-core==0.2.38
langchain-community==0.2.12
langchain-openai==0.1.22
openai==1.44.1
arxiv==2.1.3
python-dotenv==1.0.1
pydantic==2.9.2
httpx==0.27.2
numpy==1.26.4
regex==2024.9.11
pypdf==4.3.1
rapidfuzz==3.9.7
click==8.1.7
rich==13.9.2
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
